Push error handling into refactored clients
