+++ src/.vscode/settings.json
@@ -1,4 +1,4 @@
 {
-    "dotnet-test-explorer.autoExpandTree": true,
-    "dotnet-test-explorer.testProjectPath": "./SolarLunarName/SolarLunarName.Standard.Tests"
+    "dotnet-test-explorer.autoExpandTree": false,
+    "dotnet-test-explorer.testProjectPath": "./SolarLunarName/*.Tests/"
 }
